worker/requirements.txt
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
httpx==0.26.0
structlog==24.1.0
prometheus-client==0.19.0
tenacity==8.2.3
python-dateutil==2.8.2
pyyaml==6.0.1
icalendar==5.0.11
email-validator==2.1.0.post1
pydantic==2.6.1
pydantic-settings==2.1.0
aiosmtplib==3.0.1
aioimaplib==1.0.1
